Don't disable the install button permanently
The button is enabled only from onCreate(), but mOkCanInstall is
true only if the list has been fully scrolled. If mOkCanInstall is
false when onResume() is called, the button gets disabled and
never re-enabled, even if mOkCanInstall becomes true. Since the
install button can be used to scroll the list when mOkCanInstall
is false, add a separate flag so that we can re-enable the button
even if mOkCanInstall is false.

Fixes: cec631f26e2c ("DO NOT MERGE Disable overlays while installer is resumed")

REGRESSION-1262

Change-Id: I819aab95506672d291e07193b71f074292862754